GUID/UUID support

+++ driver/statement.cpp
@@ -150,9 +150,9 @@ namespace {
 
         char buf[256];
         const auto written = std::snprintf(buf, lengthof(buf), "%08x-%04x-%04x-%02x%02x-%02x%02x%02x%02x%02x%02x",
-            guid.Data1, guid.Data2, guid.Data3,
-            guid.Data4[0], guid.Data4[1], guid.Data4[2], guid.Data4[3],
-            guid.Data4[4], guid.Data4[5], guid.Data4[6], guid.Data4[7]
+            (unsigned int)guid.Data1, (unsigned int)guid.Data2, (unsigned int)guid.Data3,
+            (unsigned int)guid.Data4[0], (unsigned int)guid.Data4[1], (unsigned int)guid.Data4[2], (unsigned int)guid.Data4[3],
+            (unsigned int)guid.Data4[4], (unsigned int)guid.Data4[5], (unsigned int)guid.Data4[6], (unsigned int)guid.Data4[7]
         );
 
         if (written < 36 || written >= lengthof(buf))
